requirements.txt
# --- Brains & Formatting ---
openai
instructor
pydantic

# --- The "Eyes" (Scraping & Enrichment) ---
playwright
tavily-python

# --- Local Memory ---
chromadb

# --- Environment ---
python-dotenv
python-dateutil
